alex/arrows-label-placement: tweak arrow placement geometry (#2593)
Not for merging, just putting the PR up for the preview deploy

These are the changes from #2557. Some of them I think we should
incorporate into the original (👍) and others don't work so well (👎)
- 👍 Calculate label size & placement outside of `getGeometry`. This
fixes the issue where an empty editing text label shows up in the middle
instead of the correct position.
- 👍 Simplify label offset geometry. We can test against rectangle
polygons instead of doing a line intersection for each edge of the box
- 👎 ~~Switch 0-1 from being the ends of the arrow to the ends of the
possible range. This creates a weird slidey effect as you move the ends
of the arrow around and the possible region slides over the edge of the
box~~ i've reverted this change in 66ddb18c1
- 👍 Fix short-arrow crashes. Simpler geometry makes it a bit easier to
work out when we're in an invalid state. if the arrow is too short, we
force the label back into the center of the arrow.
- ❓ Debug geometry. I added some utilities and debug code for outputting
geometry that shows in the debug view, but isn't incorporated in
calculations. We should probably remove the specific debug code for
arrows, but keeping the library-level stuff might be worth while?

### Change Type

- [ ] `patch` — Bug fix
- [ ] `minor` — New feature
- [ ] `major` — Breaking change
- [ ] `dependencies` — Changes to package dependencies[^1]
- [ ] `documentation` — Changes to the documentation only[^2]
- [ ] `tests` — Changes to any test code only[^2]
- [ ] `internal` — Any other changes that don't affect the published
package[^2]
- [ ] I don't know

[^1]: publishes a `patch` release, for devDependencies use `internal`
[^2]: will not publish a new version

### Test Plan

1. Add a step-by-step description of how to test your PR here.
2.

- [ ] Unit Tests
- [ ] End to end tests

### Release Notes

- Add a brief release note for your PR here.